Delete dead WebSocket connections
* TTL ensured connections are deleted after 3 hours. API Gateway terminates them within 2 hours anyway
* Connections that are discovered to have been dropped (GoneException) are deleted
